1301_3.6

+++ kodilla-basic-tests/src/main/java/com/kodilla/basic_assertion/Calculator.java
@@ -1,19 +1,19 @@
 package com.kodilla.basic_assertion;
 
 public class Calculator {
-    int a;
-    int b;
+    double a;
+    double b;
 
-    public int sum(int a, int b) {
+    public double sum(double a, double b) {
         return a + b;
     }
 
-    public int subtract(int a, int b) {
+    public double subtract(double a, double b) {
 
         return a - b;
     }
 
-    public int square(int a) {
+    public double square(double a) {
         return a*a;
     }
 }
